Web/backend/requirements.txt
# Django core
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2

# ML and data processing
pandas==1.5.3
scikit-learn==1.1.3
joblib==1.2.0

# External communication
requests==2.31.0

# NLP tools
spacy==3.7.2
googletrans==4.0.0-rc1
textblob==0.17.1

# Env management
python-dotenv==1.0.0

# Deployment
gunicorn
